Functional twitch state though not ideally parameterized.
Specifically, the myocyte_model.m now saves the states as an output file from the ODE model.
The Results exploration has a significant amount more of data and twitch simulations, as well as viewing some ODE results.
Sweep has small tweeks.
Iin_interp_ca.cu currently multipllies the aficamten twitch data by ten to generate a response, but this is  not the best data necessarily.
Problem defines also reduces the number of timesteps to 3 seconds and the number of repeats.